Jwt

See merge request project-se/2122/ivse2/IVSE2-MUCKERS/pokerapp!13

+++ lib/backend/dao/user/UserDAO.ts
@@ -1,6 +1,7 @@
 import { HttpCode } from '../../entities/server/HttpCode';
 import BaseDAO, { type Response } from '../BaseDAO';
 import { type User } from '../../entities/user/User';
+import jwt from 'jsonwebtoken'
 
 class UserDAO extends BaseDAO {
 	private DEFAULT_AVATAR = 'default.png'
@@ -13,14 +14,14 @@ class UserDAO extends BaseDAO {
 
 	private updateAvatar = async(oldAvatar: string, user: Partial<User>): Promise<string> => {
 		const timePreset = Date.now()
-		const newAvatarPath = `${this.ASSET_PATH}${user.email}_${timePreset}.png`
 		const oldAvatarPath = `${this.ASSET_PATH}${oldAvatar}`
+		const newAvatarPath = `${this.ASSET_PATH}${user.email}_${timePreset}.png`
 
 		await this.writeToDisk(
 			user.profilePicture as File,
-			oldAvatarPath
+			oldAvatar !== this.DEFAULT_AVATAR ? oldAvatarPath : newAvatarPath
 		);
-		this.renameFile(oldAvatarPath, newAvatarPath)
+		if(oldAvatar !== this.DEFAULT_AVATAR) this.renameFile(oldAvatarPath, newAvatarPath)
 		return `${user.email}_${timePreset}.png`
 	}
 
@@ -42,59 +43,66 @@ class UserDAO extends BaseDAO {
 		}
 	};
 
-	public removeUser = async (email: User['email']): Promise<Response> => {
+	public removeUser = async (email: User['email'], token:string): Promise<Response> => {
 		try {
-			await this.openDbConnection();
-			const db = await this.getCollection('users');
-			const res = await db.deleteOne({ email: email });
-			await this.closeDbConnection();
-			return res.deletedCount > 0
-				? this.httpResponse(HttpCode.SUCCESS, { success: 'Deleted the user' })
-				: this.httpResponse(HttpCode.BAD_REQUEST, {
-						error: 'Something went wrong with deleting a user'
-				  });
+			if(this.verifyJWT(token, email)){
+				await this.openDbConnection();
+				const db = await this.getCollection('users');
+				const res = await db.deleteOne({ email: email });
+				await this.closeDbConnection();
+				return res.deletedCount > 0
+					? this.httpResponse(HttpCode.SUCCESS, { success: 'Deleted the user' })
+					: this.httpResponse(HttpCode.BAD_REQUEST, {
+							error: 'Something went wrong with deleting a user'
+					});
+			}else{
+				this.httpResponse(HttpCode.UNAUTHORIZED, { error: 'session compromised.' })
+			}
 		} catch (err) {
 			console.log('try catch error Register:\n', err);
 			return this.httpResponse(HttpCode.SERVER_ERROR, { error: 'User could not be deleted' });
 		}
 	};
 
-	public updateProfile = async (data: FormData): Promise<Response> => {
+	public updateProfile = async (data: FormData, token:string): Promise<Response> => {
 		try {
 			let user: User;
 			for (const pair of data.entries()) {
 				user = { ...user, ...{ [pair[0]]: pair[1] } };
 			}
-
-			await this.openDbConnection();
-			const db = await this.getCollection('users');
-			const currentUser = await db.findOne({ email: user.email }).then((result) => {
-				return result ? result : false;
-			});
-
-			if (currentUser) {
-				if (user.profilePicture) {
-					user.profilePicture = await this.updateAvatar(currentUser['profilePicture'], user)
-				}
-				const res = await db.findOneAndUpdate(
-					{ email: user.email },
-					{ $set: user },
-					{ upsert: false, returnDocument: 'after', projection: { _id: 0, password: 0 } }
-				);
-				await this.closeDbConnection();
-
-				return res
-					? this.httpResponse(HttpCode.SUCCESS, res)
-					: this.httpResponse(HttpCode.BAD_REQUEST, { error: 'could not update user' });
-			} else {
-				return this.httpResponse(HttpCode.NOT_FOUND, {
-					error: 'No user found with: ' + user.email
+			if(this.verifyJWT(token, user.email)){
+				await this.openDbConnection();
+				const db = await this.getCollection('users');
+				const currentUser = await db.findOne({ email: user.email }).then((result) => {
+					return result ? result : false;
 				});
+
+				if (currentUser) {
+					if (user.profilePicture) {
+						user.profilePicture = await this.updateAvatar(currentUser['profilePicture'], user)
+					}
+					const res = await db.findOneAndUpdate(
+						{ email: user.email },
+						{ $set: user },
+						{ upsert: false, returnDocument: 'after', projection: { _id: 0, password: 0 } }
+					);
+					await this.closeDbConnection();
+					return res
+						? this.httpResponse(HttpCode.SUCCESS, res)
+						: this.httpResponse(HttpCode.BAD_REQUEST, { error: 'could not update user' });
+				} else {
+					return this.httpResponse(HttpCode.NOT_FOUND, {
+						error: 'No user found with: ' + user.email
+					});
+				}
+			}else{
+				this.httpResponse(HttpCode.UNAUTHORIZED, { error: 'session compromised.' })
 			}
 		} catch (err) {
 			console.log('try catch error updateProfile:\n', err);
 			return this.httpResponse(HttpCode.SERVER_ERROR, { error: 'could not update user' });
 		}
+			
 	};
 
 	public register = async (user: User): Promise<Response> => {
@@ -126,23 +134,41 @@ class UserDAO extends BaseDAO {
 		try {
 			await this.openDbConnection();
 			const db = await this.getCollection('users');
+			user.password = this.hash(user.password)
+			const token = jwt.sign({email: user.email}, 'my-key', { expiresIn: "7d" })
 			const profile = await db
 				.findOne(
-					{ email: user.email, password: this.hash(user.password) },
+					{ email: user.email, password: user.password },
 					{ projection: { _id: 0, password: 0 } }
 				)
 				.then((result) => {
 					return result ? result : false;
 				});
 			await this.closeDbConnection();
+
+			const today = new Date();
+			const expire = new Date();
+			expire.setTime(today.getTime() + 3600000*24*7);
+			const cookieHeader = { 'set-cookie': `token=${token}; Path=/; expires=${expire} ;HttpOnly` }
+
 			return profile
-				? this.httpResponse(HttpCode.SUCCESS, profile)
+				? this.httpResponse(HttpCode.SUCCESS, profile , cookieHeader)
 				: this.httpResponse(HttpCode.NOT_FOUND, { error: 'Wrong credentials' });
 		} catch (err) {
 			console.log('try catch error Login:\n', err);
 			return this.httpResponse(HttpCode.SERVER_ERROR, { error: 'Could not login' });
 		}
 	};
+ 
+	public logout = async (): Promise<Response> => {
+		try{
+			const cookieHeader = { 'set-cookie': `token=deleted; Path=/; expires=Thu, 01 Jan 1970 00:00:00 GMT` }
+			return this.httpResponse(HttpCode.SUCCESS, {} ,cookieHeader)
+		}catch(err){
+			console.log('try catch error Logout:\n', err);
+			return this.httpResponse(HttpCode.SERVER_ERROR, { error: 'Could not logout' });
+		}
+	}
 }
 
 export default UserDAO;
